Keras/requirements.txt
numpy==1.12.1
scipy==0.19.0
h5py==2.7.0
tensorflow==1.5.0
keras==2.1.4
flask==0.12.2
gevent==1.2.1
imutils==0.4.5
requests==2.14.2
redis==2.10.6
Pillow==4.1.1
